Move each_top_level_statement and readmultiline to Irb
This will save RubyLex from knowning information about commands and aliases.
